requirements.txt
absl-py==1.2.0
aiohttp==3.8.6
aiosignal==1.3.1
altair==4.2.0
annotated-types==0.6.0
antiorm==1.2.1
anyio==3.7.1
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
blinker==1.5
bs4==0.0.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
cinemagoer==2022.2.11
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cycler==0.11.0
dataclasses-json==0.6.1
db==0.1.1
db-sqlite3==0.0.1
debugpy==1.6.3
decorator==5.1.1
distlib==0.3.6
emoji==2.0.0
entrypoints==0.4
exceptiongroup==1.1.3
executing==0.10.0
filelock==3.8.0
Flask==2.1.0
flatbuffers==1.12
fonttools==4.36.0
frozenlist==1.4.0
fsspec==2023.10.0
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.47.0
gunicorn==20.0.4
h11==0.14.0
h5py==3.7.0
htbuilder==0.6.1
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.18.0
idna==3.3
IMDbPY==2022.7.9
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
ipywidgets==8.0.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.13.0
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-widgets==3.0.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langchain==0.0.326
langsmith==0.0.54
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.20.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
more-itertools==9.1.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.5
nltk==3.5
numpy==1.23.2
oauthlib==3.2.0
openai==0.28.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.30
protobuf==4.24.4
psutil==5.9.1
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
PyYAML==6.0
pyzmq==23.2.1
regex==2022.8.17
replicate==0.15.5
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
scipy==1.9.0
seaborn==0.11.2
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
stack-data==0.4.0
streamlit==1.28.0
streamlit-chat==0.1.1
tenacity==8.2.3
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-cpu==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2022.2
tzlocal==4.2
uritools==4.0.0
urlextract==1.6.0
urllib3==1.26.11
validators==0.20.0
virtualenv==20.21.0
watchdog==2.1.9
wcwidth==0.2.5
Werkzeug==2.2.2
widgetsnbextension==4.0.2
wordcloud==1.8.2.2
wrapt==1.14.1
yarl==1.9.2
zipp==3.8.1
